requirements.txt
pyaudio
SpeechRecognition
--extra-index-url https://download.pytorch.org/whl/cu116
torch
numpy
git+https://github.com/openai/whisper.git
python-dotenv
python-youtube
wolframalpha
requests
giphy_client
tenacity
openai
pika